Swap array order to enable 'Page specific' as default
Swap the array order to enable 'Page specific' as the default setting when editing post or page. The use case is a large blog with multiple authors, not everyone remembers or knows to set the default 'Completely' to 'Page specific' when editing a published post or page.  The result is that a cache with thousands of pages that has been pre-warmed (using curl against a sitemap of URIs) can get inadvertently wiped out.  I am proposing that the default be 'Page specific' instead.  Thanks!