Cache http and https pages separately
A single URI may be served as http or as https content.

In this case, both requests are likely to not serve the same content.
The first one will contain many "http" links to website resources
whereas the second will point them as https links.

Before this commit, a page cache was created the first time a URI
was requested, and all next requests to this URI were answered
with cached content, ignoring scheme. This especially led to
"mixed content" issues when the cache contained http content,
because a https page was answered with http links to resources.

This commit inserts web-server TCP port in cached page file name
so http (likely port 80) and https (likely port 443) are saved
as distinct files.

Notes:

Injecting SERVER_PORT instead of a "http"/"https" token has been choosen
for IIS compatibility, where detecting https appears always weak.

Injecting it in file name has been choosen to easily delete all of them
when clearing a single-page cache.