add scheme-based caching (#94)
Cache HTTP and HTTPS pages separately (#20). This builds upon PR #50.
This fixes HTTP resources being requested by a page that is loaded over
HTTPS. This occurs when a page is loaded over HTTP, cached, and then
later loaded over HTTPS. This also fixes HTTPS resources being
requested by a page that is loaded over HTTP. If the website does not
have a valid SSL/TLS certificate the requests would fail. This issue
can arise when HTTP or HTTPS is not being forced.

Update Cache Enabler HTML comment to be more clear on what cached page
is being delivered.

Update `clear_home()` method to unlink based on glob pattern. This
builds upon part of PR #64. This is a better way to clear the home page
cache as we are not attempting to unlink files that do not exist.
Further, this makes clearing the home page cache more simple when
introducing scheme-based caching.

Fixes #20