update clear requests and fix advcache
Update `Cache_Enabler::process_clear_request()` to only clear the home
page cache when using the "Clear URL Cache" admin bar button on the
home page. Previously the complete cache would be cleared.

Update how `Cache_Enabler_Disk::_file_scheme()` returns `https` or
`http`. This is required to return the correct scheme if the WordPress
installation is behind a proxy.

Fix advanced cache path to cached variants. This allows the advanced
cache to deliver the cached page and bypass PHP as it did in version
1.3.5. This should have been done in PR #94.

Delete `CE_SETTINGS_PATH` constant that was added in PR #96. The
advanced cache is not always able to read the value of this constant.