fix cache clearing (#103)
Fix cache clearing issue for subdirectory multisite networks that was
introduced in version 1.4.0 when I added
`Cache_Enabler::clear_blog_id_cache()` (PR #96). This caused only the
main site in a subdirectory multisite network to be cleared properly
when using the "Clear Cache" admin bar button. This occurred because
the blog path was duplicated in `$clear_url`.