add cache bypass method for sitemaps
Add cache bypass method for sitemaps by checking the file extension. (I
did not find a requirement to verify the content type.) When checking
for a sitemap in particular I believe this method would be the fastest
and will cover both core (since WordPress version 5.5) and third party
sitemaps.

Fixes #101