update getting wp-config.php
Update where the `wp-config.php` file can be found. Use the same logic
that the WordPress core uses to load the `wp-config.php` file in
`wp-load.php` (as of version 5.5). This extends the search to find the
`wp-config.php` file if it has been moved one directory level above the
WordPress installation and is not part of another installation.

Closes #105