fix advcache settings for subdirectory networks
Fix advanced cache settings recognition when on the home page of the
main site in subdirectory networks. Previously the main site advanced
cache settings file was not being found when on the home page of the
main site because the blog path was being returned as `-`. This was not
finding the main site advanced cache settings file when on the home
page because searching for `-` and trying to replace it in the settings
file path to find the main site advanced cache settings will not work
(because there are many `-` in the settings file path). Instead, ensure
an empty blog path is returned when on the home page of the main site.