update getting file path
In the advanced cache file remove the `$path` parameter from the
`_ce_file_path()` function because it is unnecessary and then rename it
to `_file_path()` because it copies `Cache_Enabler_Disk::_file_path()`.
(This is how our other functions in this file are named.) Simplify the
`is_file()` check in the advanced cache `_file_path()` and
`Cache_Enabler_Disk::_file_path()` because it returns a boolean.