fix trailing slash handling
Add `Cache_Enabler::_is_trailing_slash_redirect()` (mostly a copy of
`Cache_Enabler::_bypass_cache_for_trailing_slash()` that was removed in
PR #100 due to my mistake. I am unsure how the tests I performed lead
me to believe checking the trailing slash was no longer required. It
turns out we receive a `200` status code, which is why in this case
checking the status will not allow the trailing slash redirect to
occur. Complete cache should still be cleared if the permalink
structure is changed because this can update the cache structure.

`self::_is_trailing_slash_redirect()` will come after
`self::_is_sitemap()` in the conditional tag check until the regex can
be improved to not count a file with the extension as an asset that is
missing a trailing slash when set (e.g. `wp-sitemap.xml`).