fix getting URL for subdirectory installations
Update `Cache_Enabler::clear_home_page_cache()`,
`Cache_Enabler::clear_blog_id_cache()`, and
`Cache_Enabler::get_blog_domain()` to use `get_home_url()`, which gets
the Site Address (URL) setting, instead of `get_site_url()`, which gets
the WordPress Address (URL) setting. This fixes the issue where the
wrong URL is retrieved if the WordPress installation is installed in a
subdirectory and the Site Address (URL) setting is different from the
WordPress Address (URL) setting. This is required to ensure the URL
retrieved will be what visitors receive on the front end if the two
settings are different, because the Site Address (URL) setting
determines the cache structure.

Fixes #113