fix WebP URL conversion
Previously the entire full match string was being converted to
lowercase to check the image count. This was introduced in version
1.4.8 (PR #116). This was not the best way to handle this as it could
result in `404` errors in some cases (depending on the server
configuration). We should not be changing the entire URLs. Instead, we
should only be appending the `.webp` extension or replacing the default
extension with `.webp`. Use `preg_match_all()` with a case insensitive
pattern to get the image count instead of `substr_count()`. This will
be a better solution than transforming the `$full_match` to lowercase
before each instance of `substr_count`.

Fixes #119