update clear request processing
Update clear request processing:

* Redirect to same page even when in an admin interface. This will
prevent the clear request query string from staying in the browser
address bar for a cleaner way of clearing the cache. This means the
clear ID is no longer required and has been removed.

* Display notice in admin interface page by checking for the
`_cache_enabler_cache_cleared_<uid>` transient and then deleting it
after the notice has been shown. This was done instead of adding a
query parameter, like `_cache-cleared`, and then removing it after the
page has loaded by including it in the `removable_query_args` filter
(to then be removed by `wp_admin_canonical_url()`). This was done
because it allows the notice to be displayed in any case needed, like
after the cache has been cleared when saving the Cache Enabler
settings. Furthermore, this behavior was not copied from how WordPress
displays certain admin notices because first the redirect occurs and
then the query parameter is removed, whereas this way it just redirects
to the same page.

* Update notice messages to "Cache cleared." and "Network cache
cleared." instead of "The cache has been cleared.".

Update notices being called with `admin_notices` to not use the
`show_message()` function as this is unnecessary when printing admin
notices from this hook.

Combine `Cache_Enabler::warning_is_permalink()` with
`Cache_Enabler::requirements_check()` as it is also a requirement of
Cache Enabler.

Add `Cache_Enabler::_is_plain_permalink_structure()` in place of
`Cache_Enabler_Disk::is_permalink()`. Bypass the cache if a plain
permalink structure is set instead of only creating an instance of
`Cache_Enabler_Disk` if a custom permalink structure is set.