fix creating settings file
Fix creating settings file on upgrade hook. Like when in the activation
hook, the `update_option_cache_enabler` action hook is not fired when
in the upgrade hook. This means even though the database was being
updated, the hook required to create the settings file was not being
fired. Instead, check if the `update_option_cache_enabler` hook has
been fired. If it has not create the settings file. This will now cover
creating the settings file when this method is called by the upgrade
hook and continue covering if the settings file should be force
created/updated even if the database did not need to be updated.
(WordPress will not update the database if the same value is trying to
be added.)