update cache engine
Update cache delivery to not check if the cache has expired as this is
already done previously in the stack (when the engine is starting).
Having this additional check is not currently required with having the
cache delivery only called in the `advanced-cache.php` drop-in.

Update late cache engine start to be on the `init` hook instead of
`plugins_loaded` when the plugin is initialized (#152).

Add cache engine start check to see whether or not the cache engine
should be started, moving checking if the request is an Ajax, REST API,
or XMLRPC to this method instead to prevent unnecessary engine starts.
This fixes the undefined index notice errors when using any WP-CLI
command (#152).

Add deprecated variable that was previously deleted in version 1.5.0 to
improve backwards compatibility
(nosilver4u/ewww-image-optimizer#199).

Fixes #152